connection with discord rpc

+++ src/H3MP/Extensions/DateTimeExtensions.cs
@@ -4,9 +4,9 @@ namespace H3MP.Extensions
 {
 	public static class DateTimeExtensions
 	{
-		public static long ToUnixTimestamp(this DateTime @this)
+		public static int ToUnixTimestamp(this DateTime @this)
 		{
-			return (long) @this.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
+			return (int) @this.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
 		}
 	}
 }
